Change to URL of the GDS distribution
In order to comply with the sanctions against Russia,
we are moving the GDS artifacts to a new bucket that sits
behind Cloudflare and Route53 with a new domain.

Unfortunately, we cannot configure redirects from the current S3
bucket, because redirects only apply if the bucket is configured
to use static website hosting, but we used the direct object URLs.
Regardless of bucket configuration, using the direct object URL
on a public object will not issue a redirect.